Final Commit
Ready for Review. Set the date to the current date. Added in the call of the calDays function. Adjusted the daysInMonth function to keep track of leap years. Defined the calDays function. Made the calDays function return the table rows for the calendar. Highlighted the current day. Added events to the calendar.